+++ facet-msgpack/Cargo.toml
@@ -18,6 +18,7 @@ categories = ["encoding", "parsing", "data-structures"]
 
 [dependencies]
 facet-trait = { workspace = true }
+facet-types.workspace = true
 [dev-dependencies]
 facet = { workspace = true }
 facet-derive = { workspace = true }
