+++ facet-msgpack/src/lib.rs
@@ -1,230 +1,17 @@
 #![warn(missing_docs)]
 #![doc = include_str!("../README.md")]
 
-use facet_core::Partial;
-use std::convert::TryInto;
-
 mod errors;
-use errors::Error as DecodeError;
+pub use errors::Error as DecodeError;
 
 mod constants;
-use constants::*;
+pub use constants::*;
 
 #[cfg(test)]
 mod tests;
 
-/// Deserializes MessagePack-encoded data into a Facet Partial.
-///
-/// This function takes a MessagePack byte array and populates a Partial object
-/// according to the shape description provided by the Partial.
-///
-/// # Example
-///
-/// ```
-/// use facet::Facet;
-/// use facet_msgpack::from_msgpack;
-///
-/// #[derive(Debug, Facet, PartialEq)]
-/// struct User {
-///     id: u64,
-///     username: String,
-/// }
-///
-/// // MessagePack binary data (equivalent to {"id": 42, "username": "user123"})
-/// let msgpack_data = [
-///     0x82, 0xa2, 0x69, 0x64, 0x2a, 0xa8, 0x75, 0x73,
-///     0x65, 0x72, 0x6e, 0x61, 0x6d, 0x65, 0xa7, 0x75,
-///     0x73, 0x65, 0x72, 0x31, 0x32, 0x33
-/// ];
-///
-/// let mut partial = User::partial();
-/// from_msgpack(&mut partial, &msgpack_data).expect("Failed to parse MessagePack data");
-///
-/// let user = partial.build::<User>();
-/// assert_eq!(user, User { id: 42, username: "user123".to_string() });
-/// ```
-///
-/// # Parameters
-/// * `partial` - A mutable reference to a Partial object that will be filled with deserialized data
-/// * `msgpack` - A byte slice containing MessagePack-encoded data
-///
-/// # Returns
-/// * `Ok(())` if deserialization was successful
-/// * `Err(DecodeError)` if an error occurred during deserialization
-///
-/// # MessagePack Format
-/// This implementation follows the MessagePack specification:
-/// <https://github.com/msgpack/msgpack/blob/master/spec.md>
-#[allow(clippy::needless_lifetimes)]
-pub fn from_msgpack(partial: &mut Partial, msgpack: &[u8]) -> Result<(), DecodeError> {
-    let mut decoder = Decoder::new(msgpack);
+mod from_msgpack;
+pub use from_msgpack::*;
 
-    fn deserialize_value(decoder: &mut Decoder, partial: &mut Partial) -> Result<(), DecodeError> {
-        let shape_fn = partial.shape();
-        let shape = shape_fn.get();
-
-        match &shape.innards {
-            facet_core::Def::Scalar(scalar) => {
-                let slot = partial.scalar_slot().expect("Scalar slot");
-                match scalar {
-                    facet_core::Scalar::String => {
-                        let value = decoder.decode_string()?;
-                        slot.fill(value);
-                    }
-                    facet_core::Scalar::U64 => {
-                        let value = decoder.decode_u64()?;
-                        slot.fill(value);
-                    }
-                    _ => {
-                        println!("Unsupported scalar type: {:?}", scalar);
-                        todo!()
-                    }
-                }
-            }
-            facet_core::Def::Struct { .. } => {
-                let map_len = decoder.decode_map_len()?;
-
-                for _ in 0..map_len {
-                    let key = decoder.decode_string()?;
-                    let slot = partial
-                        .slot_by_name(&key)
-                        .map_err(|_| DecodeError::UnknownField(key))?;
-
-                    let mut partial_field = Partial::alloc(slot.shape());
-                    deserialize_value(decoder, &mut partial_field)?;
-                    slot.fill_from_partial(partial_field);
-                }
-            }
-            _ => {
-                println!("Unsupported shape: {:?}", shape.innards);
-                todo!()
-            }
-        }
-
-        Ok(())
-    }
-
-    deserialize_value(&mut decoder, partial)
-}
-
-struct Decoder<'input> {
-    input: &'input [u8],
-    offset: usize,
-}
-
-impl<'input> Decoder<'input> {
-    fn new(input: &'input [u8]) -> Self {
-        Decoder { input, offset: 0 }
-    }
-
-    /// Decodes a single byte from the input.
-    /// This is a low-level method used by other decoders.
-    fn decode_u8(&mut self) -> Result<u8, DecodeError> {
-        if self.offset >= self.input.len() {
-            return Err(DecodeError::InsufficientData);
-        }
-        let value = self.input[self.offset];
-        self.offset += 1;
-        Ok(value)
-    }
-
-    /// Decodes a 16-bit unsigned integer in big-endian byte order.
-    /// This is a low-level method used by other decoders.
-    fn decode_u16(&mut self) -> Result<u16, DecodeError> {
-        if self.offset + 2 > self.input.len() {
-            return Err(DecodeError::InsufficientData);
-        }
-        let value =
-            u16::from_be_bytes(self.input[self.offset..self.offset + 2].try_into().unwrap());
-        self.offset += 2;
-        Ok(value)
-    }
-
-    /// Decodes a 32-bit unsigned integer in big-endian byte order.
-    /// This is a low-level method used by other decoders.
-    fn decode_u32(&mut self) -> Result<u32, DecodeError> {
-        if self.offset + 4 > self.input.len() {
-            return Err(DecodeError::InsufficientData);
-        }
-        let value =
-            u32::from_be_bytes(self.input[self.offset..self.offset + 4].try_into().unwrap());
-        self.offset += 4;
-        Ok(value)
-    }
-
-    /// Decodes a MessagePack-encoded unsigned 64-bit integer.
-    /// Handles the following MessagePack types:
-    /// - positive fixint (0x00 - 0x7f): single-byte positive integer
-    /// - uint8 (0xcc): 8-bit unsigned integer
-    /// - uint16 (0xcd): 16-bit unsigned integer (big-endian)
-    /// - uint32 (0xce): 32-bit unsigned integer (big-endian)
-    /// - uint64 (0xcf): 64-bit unsigned integer (big-endian)
-    ///
-    /// Ref: <https://github.com/msgpack/msgpack/blob/master/spec.md#int-format-family>
-    fn decode_u64(&mut self) -> Result<u64, DecodeError> {
-        match self.decode_u8()? {
-            MSGPACK_UINT8 => Ok(self.decode_u8()? as u64),
-            MSGPACK_UINT16 => Ok(self.decode_u16()? as u64),
-            MSGPACK_UINT32 => Ok(self.decode_u32()? as u64),
-            MSGPACK_UINT64 => {
-                if self.offset + 8 > self.input.len() {
-                    return Err(DecodeError::InsufficientData);
-                }
-                let value = u64::from_be_bytes(
-                    self.input[self.offset..self.offset + 8].try_into().unwrap(),
-                );
-                self.offset += 8;
-                Ok(value)
-            }
-            prefix @ MSGPACK_POSFIXINT_MIN..=MSGPACK_POSFIXINT_MAX => Ok(prefix as u64),
-            _ => Err(DecodeError::UnexpectedType),
-        }
-    }
-
-    /// Decodes a MessagePack-encoded string.
-    /// Handles the following MessagePack types:
-    /// - fixstr (0xa0 - 0xbf): string up to 31 bytes
-    /// - str8 (0xd9): string up to 255 bytes
-    /// - str16 (0xda): string up to 65535 bytes
-    /// - str32 (0xdb): string up to 4294967295 bytes
-    ///
-    /// Ref: <https://github.com/msgpack/msgpack/blob/master/spec.md#formats-str>
-    fn decode_string(&mut self) -> Result<String, DecodeError> {
-        let prefix = self.decode_u8()?;
-
-        let len = match prefix {
-            prefix @ MSGPACK_FIXSTR_MIN..=MSGPACK_FIXSTR_MAX => (prefix & 0x1f) as usize,
-            MSGPACK_STR8 => self.decode_u8()? as usize,
-            MSGPACK_STR16 => self.decode_u16()? as usize,
-            MSGPACK_STR32 => self.decode_u32()? as usize,
-            _ => return Err(DecodeError::UnexpectedType),
-        };
-
-        if self.offset + len > self.input.len() {
-            return Err(DecodeError::InsufficientData);
-        }
-
-        let value = String::from_utf8(self.input[self.offset..self.offset + len].to_vec())
-            .map_err(|_| DecodeError::InvalidData)?;
-        self.offset += len;
-        Ok(value)
-    }
-
-    /// Decodes a MessagePack-encoded map length.
-    /// Handles the following MessagePack types:
-    /// - fixmap (0x80 - 0x8f): map with up to 15 elements
-    /// - map16 (0xde): map with up to 65535 elements
-    /// - map32 (0xdf): map with up to 4294967295 elements
-    ///
-    /// Ref: <https://github.com/msgpack/msgpack/blob/master/spec.md#formats-map>
-    fn decode_map_len(&mut self) -> Result<usize, DecodeError> {
-        let prefix = self.decode_u8()?;
-
-        match prefix {
-            prefix @ MSGPACK_FIXMAP_MIN..=MSGPACK_FIXMAP_MAX => Ok((prefix & 0x0f) as usize),
-            MSGPACK_MAP16 => Ok(self.decode_u16()? as usize),
-            MSGPACK_MAP32 => Ok(self.decode_u32()? as usize),
-            _ => Err(DecodeError::UnexpectedType),
-        }
-    }
-}
+mod to_msgpack;
+pub use to_msgpack::*;
