+++ facet-msgpack/src/to_msgpack.rs
@@ -1,0 +1,3 @@
+pub fn to_vec<T>(_t: &T) -> Result<Vec<u8>, ()> {
+    todo!()
+}
